Users can now save realtime blackboards as silent animations
