+++ src/components/CreateExplanation.vue
@@ -17,7 +17,7 @@
       <div v-if="(newExplanationDbRef || postDbRef)" class="d-flex align-center">
         <template v-if="user">
           <v-btn v-if="!isUploadingPost"
-            @click="submitExplanation()" 
+            @click="uploadExplanation()" 
             :loading="isButtonDisabled" 
             :disabled="isButtonDisabled"
             color="secondary" 
@@ -91,6 +91,7 @@
 <script>
 import Blackboard from "@/components/Blackboard.vue";
 import DatabaseHelpersMixin from "@/mixins/DatabaseHelpersMixin.js";
+import ExplUploadHelpers from "@/mixins/ExplUploadHelpers.js";
 import DoodleVideo from "@/components/DoodleVideo.vue";
 import TextEditor from "@/components/TextEditor.vue";
 import BasePopupButton from "@/components/BasePopupButton.vue";
@@ -116,7 +117,8 @@ export default {
     newExplanationDbRef: Object
   },
   mixins: [
-    DatabaseHelpersMixin
+    DatabaseHelpersMixin,
+    ExplUploadHelpers
   ],
   components: { 
     Blackboard, 
@@ -177,10 +179,10 @@ export default {
       const { BlackboardDrawingCanvas } = Blackboard.$refs; 
       BlackboardDrawingCanvas.resizeBlackboard();
     },
+    // The 2 methods below are used by `NewPost` and `SeePost`
     getBlackboard () {
       return this.$refs.Blackboard;
     },
-    // used by NewExplanation and SeeExplanation
     getTextEditor () {
       return this.$refs.TextEditor;
     },
@@ -197,128 +199,19 @@ export default {
       this.isRecordingVideo = false;
       this.isPreviewing = true;
     },
-    /**
-     * Save the explanation data into the global Vuex store (so the data doesn't disappear
-     * if this component is destroyed), then initiates the upload process.
-     */
-    async submitExplanation () {
+    async uploadExplanation() {
       const { TextEditor, Blackboard } = this.$refs;
       if (this.postTitle.length === 0 && this.titleRequired) {
-        this.postTitle = `Untitled (${new Date().toDateString()})`; 
+        this.postTitle = `Untitled (${new Date().toLocaleTimeString()})`;
       }
-      const explRef = this.willCreateNewPost ? this.postDbRef : this.newExplanationDbRef.doc(getRandomId());
-      this.$store.commit("ADD_EXPL_TO_CACHE", {
-        ref: explRef, // to uniquely identify each explanation when there are simultaneous uploads
-        strokesArray: this.strokesArray,
-        backgroundImageBlob: Blackboard.getImageBlob(),
-        thumbnailBlob: this.previewVideo.thumbnailBlob ? this.previewVideo.thumbnailBlob : await Blackboard.getThumbnail(),
-        audioBlob: Blackboard.currentState === RecordState.POST_RECORD ? this.previewVideo.audio.blob : null,
-        metadata: {
-          title: this.postTitle,
-          html: TextEditor.html,
-          date: new Date().toISOString(),
-          creator: this.isAnonymous ? this.anonymousUser : this.simplifiedUser,
-          mitClass: this.mitClass,
-          tags: [],
-          duration: Blackboard.currentTime,
-          hasStrokes: this.strokesArray.length > 0
-        }
-      });
-      this.uploadExplanation(explRef);
-      this.$root.$emit("show-snackbar", "Uploading your explanation...");
-      this.$emit("upload-started"); // let the clients i.e. ClassPageSeePost and ClassPageNewPost re-render the <CreateExplanation/> component
-    },
-    /**
-     * 
-     * @param {*} context a reference to the Vuex store, and is automatically injected when the method is called using `$store.dispatch()`
-     * @param {*} ref the location on Firestore to which the explanation document will be uploaded. 
-     *                Moreover, ref.id maps to the specific explanation from the expl cache. 
-     * @returns Uploads the explanation. If the upload fails for any reason,
-     * this function will call itself after 5 seconds to re-attempt the upload
-     */
-    async uploadExplanation (ref, retryCount = 0) {
-      try {
-        const explData = this.$store.state.explCache[ref.id];
-        const { strokesArray, audioBlob, thumbnailBlob, backgroundImageBlob } = explData; 
-        const explDoc = { ...explData.metadata }; // we build up each property of `explDoc` then upload it Firestore
-        const promises = []; // upload thumbnail, audio, images and strokes in parallel 
-        if (strokesArray.length > 0 || backgroundImageBlob) { // if the blackboard was used
-          promises.push(
-            this.uploadStrokesToDatabase(strokesArray, ref.collection("strokes"))
-          );
-          promises.push(
-            this.$_saveToStorage(getRandomId(), thumbnailBlob).then(URL => explDoc.thumbnail = URL)
-          );
-          if (audioBlob) {
-            promises.push(
-              this.$_saveToStorage(getRandomId(), audioBlob).then(URL => explDoc.audioUrl = URL)
-            );
-          } 
-          if (backgroundImageBlob) {
-            promises.push(
-              this.$_saveToStorage(getRandomId(), backgroundImageBlob).then(URL => explDoc.imageUrl = URL)
-            );
-          }
-        }
-        await Promise.all(promises);
-        // TODO: refactor the below logic to the client
-        if (this.willCreateNewPost) {
-          explDoc.participants = [this.simplifiedUser];
-          explDoc.hasReplies = false;
-        } else {
-          promises.push(
-            this.postDbRef.update({
-              participants: firebase.firestore.FieldValue.arrayUnion(this.simplifiedUser),
-              hasReplies: true
-            })
-          );
-        }
-        promises.push(ref.set(explDoc));
-        await Promise.all(promises);
-        delete this.$store.state.explCache[ref.id];
-        this.$root.$emit("show-snackbar", "Successfully uploaded your explanation.");   
-      } catch (error) {
-        // TODO: send an error email to ExplainMIT core team
-        console.log("error =", error);
-        this.$root.$emit("show-snackbar", "Upload failed, trying again...");
-        const maxRetryCount = 5;
-        if (retryCount < maxRetryCount) {
-          setTimeout(() => this.uploadExplanation(ref, retryCount + 1), 5000); // set a delay in case the upload failure is immediate and will overwhelm the call stack
-        }
-      }
-    },
-    /**
-     * @param strokesArray an array of stroke objects
-     * @param databaseRef location on Firestore where the stroke documents will be uploaded
-     * @effect uploads each stroke of the array to databaseRef 
-     */
-    uploadStrokesToDatabase (strokesArray, databaseRef) {
-      return new Promise(async (resolve, reject) => {
-        try {
-          const promises = [];
-          const n = strokesArray.length;
-          let currentBatch = db.batch();
-          let currentBatchSize = 0; 
-          const maxBatchSize = 500; 
-
-          for (const stroke of this.strokesArray) {
-            if (currentBatchSize >= maxBatchSize) {
-              promises.push(currentBatch.commit());
-              currentBatch = db.batch(); 
-              currentBatchSize = 0; 
-            } 
-            currentBatch.set(databaseRef.doc(getRandomId()), stroke);
-            currentBatchSize += 1;
-          }
-
-          promises.push(currentBatch.commit()); 
-          await Promise.all(promises);
-          resolve();
-        } catch (reason) {
-          console.log("failed, reason =", reason);
-          reject(reason);
-        }
-      });
+      const thumbnailBlob = this.previewVideo.thumbnailBlob ? this.previewVideo.thumbnailBlob : await Blackboard.getThumbnail();
+      this.$_saveExplToCacheThenUpload(
+        thumbnailBlob,
+        Blackboard.currentState === RecordState.POST_RECORD ? this.previewVideo.audio.blob : null,
+        TextEditor.html,
+        this.postTitle,
+        this.willCreateNewPost ? this.postDbRef : this.newExplanationDbRef.doc(getRandomId())
+      )
     },
     toggleFullscreenDoodle () {
       this.isFullScreenDoodle = !this.isFullScreenDoodle;
